+++ changelog.md
@@ -78,10 +78,15 @@ This now needs to be written as:
   via the new ``nim doc0`` command.
 - Added ``system.getStackTraceEntries`` that allows you to access the stack
   trace in a structured manner without string parsing.
+- Added ``sequtils.mapLiterals`` for easier construction of array and tuple
+  literals.
+- Added ``parseutils.parseSaturatedNatural``.
+- ``atomic`` and ``generic`` are no longer keywords in Nim. ``generic`` used to be
+  an alias for ``concept``, ``atomic`` was not used for anything.
 - Moved from stdlib into Nimble packages:
   - [``basic2d``](https://github.com/nim-lang/basic2d)
     _deprecated: use ``glm``, ``arraymancer``, ``neo``, or another package instead_
-  - [``basic3d``](https://github.com/nim-lang/basic3d) 
+  - [``basic3d``](https://github.com/nim-lang/basic3d)
     _deprecated: use ``glm``, ``arraymancer``, ``neo``, or another package instead_
   - [``gentabs``](https://github.com/lcrees/gentabs)
   - [``libuv``](https://github.com/lcrees/libuv)
@@ -89,3 +94,7 @@ This now needs to be written as:
   - [``poly``](https://github.com/lcrees/polynumeric)
   - [``pdcurses``](https://github.com/lcrees/pdcurses)
   - [``romans``](https://github.com/lcrees/romans)
+
+- Added ``system.runnableExamples`` to make examples in Nim's documentation easier
+  to write and test. The examples are tested as the last step of
+  ``nim doc``.